refactor: change test framework narrative(gen/ext)
This commit and the next one aims to change the proposed narrative
for the new tests architecture for the library.

The new narrative comes from the idea that we have external
estimators (previously called generators, into the references
    directory). We want to compare and test those external estimators
against our internal estimators (previously declared into the
    validations directory).

This new narrative is necessary as the abstraction of the test logic
makes meaningless the previous `test_`-prefixed names of the
files/functions in `validations/`.

In this commit, I have done:

1. Change /test directory structure:
  - References -> external_estimators
  - Validations -> internal_estimators

2. Remove unnecessary nesting in the external_estimators.
  - One file per estimator should be enough.

3. Remove unnecessary name redundancy in functions.

4. Change name prefixes:
  - gen -> ext
  - test -> int

5. Improve general naming and propagate the required changes.

6. General adjust of the logic in the `generate_references.py` file.