refactor: drastic improvement on sage import func
This commit introduces a drastic improvement on the sage import function
used as helper for the `generate_kat.py` function.

The function is now able to return modules/functions/variables from a
`.sage` file as python objects, or to include them in the caller namespace,
similar to the regular Python import mechanism.