refactor: porting base changes from draft PR 161
This commit squash all the commits made in

"Refactor/new regression tests architecture #161",

Just as a way to compile all the work made there and use it
as the start point to build another (improved) framework, without
all the noise provided by more than 20 commits and a 40+
conversation.

In case of reference about any particular change, you can go to that
draft PR and see the discussion and motivations of the code produced/modified
in this commit.