feat: integrate docker-pytest in the Makefile
- Refactoring of the `make docker-test` and `make docker-testfast`
commands in the Makefile, so we are able to disable the Sage doctest
execution in whole directories. This allows us to made the transition
to python doctests in a safe way, because every doctest will be executed
either by Sage with the `docker-test` command  or by Python with
the new `docker-pytest` command.

- Deprecation of the previous `docker-pytest` as it wasn't used by our
CI, doesn't use pure python, and wasn't well-defined in its scopes (involving
sage testings, for example).

- Introduction of the new `docker-pytest` command to run pytest with pure
python, in the directories that are progressively disabled in the `docker-test`
command.